# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateObjectionDto": {
        "properties": {
          "reason": {
            "description": "Optional reason for objecting. Strongly encouraged for productive deliberation.",
            "maxLength": 1000,
            "type": "string"
          }
        },
        "type": "object"
      },
      "CreatePaymentIntentDto": {
        "properties": {
          "addonSlugs": {
            "description": "Optional add-on slugs purchased alongside the tier.",
            "example": [
              "ai_vibe_pack",
              "holiday_scanner"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "callbackUrl": {
            "description": "Where Paystack redirects after hosted-checkout completes. Defaults to WEB_ORIGIN/events/:id/return.",
            "type": "string"
          },
          "eventId": {
            "description": "Draft event the payment provisions.",
            "type": "string"
          },
          "tierSlug": {
            "description": "Slug of the chosen tier.",
            "example": "occasion",
            "type": "string"
          }
        },
        "required": [
          "eventId",
          "tierSlug"
        ],
        "type": "object"
      },
...
      "EventRefDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "scheduledDate": {
            "example": "2026-07-25",
            "nullable": true,
            "type": "string"
          },
          "title": {
            "example": "Bola's surprise 30th",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title"
        ],
        "type": "object"
      },
...
      "EventReviewDto": {
        "properties": {
          "comment": {
            "description": "Free-form comment. Surfaced only when the viewer is the reviewee or has admin rights on the event — null otherwise.",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-25T10:14:00.000Z",
            "type": "string"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "revieweeId": {
            "example": "usr_creator...",
            "type": "string"
          },
          "reviewerId": {
            "example": "usr_2N9k8x...",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "reviewerId",
          "revieweeId",
          "rating",
          "createdAt"
        ],
        "type": "object"
      },
      "EventReviewsListResponseDto": {
        "properties": {
          "isPrivileged": {
            "description": "Whether the viewer can see all reviews (admin/creator) vs only their own row. Lets the UI decide whether to render the full list or just the \"you reviewed\" card.",
            "example": true,
            "type": "boolean"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/EventReviewWithReviewerDto"
            },
            "type": "array"
          }
        },
        "required": [
          "reviews",
          "isPrivileged"
        ],
        "type": "object"
      },
      "EventReviewWithReviewerDto": {
        "properties": {
          "comment": {
            "description": "Free-form comment. Surfaced only when the viewer is the reviewee or has admin rights on the event — null otherwise.",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-25T10:14:00.000Z",
            "type": "string"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "revieweeId": {
            "example": "usr_creator...",
            "type": "string"
          },
          "reviewer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewerSummaryDto"
              }
            ],
            "description": "Reviewer profile snippet — null for non-privileged viewers.",
            "nullable": true
          },
          "reviewerId": {
            "example": "usr_2N9k8x...",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "reviewerId",
          "revieweeId",
          "rating",
          "createdAt"
        ],
        "type": "object"
      },
...
      "ObjectionRaisedByDto": {
        "properties": {
          "email": {
            "example": "bola@example.com",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "example": "usr_3K9p1y2Mn8oW5dC7eUj4tZ",
            "type": "string"
          },
          "name": {
            "example": "Bola Adeyemi",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "ObjectionResponseDto": {
        "properties": {
          "createdAt": {
            "example": "2026-05-27T09:01:14.000Z",
            "type": "string"
          },
          "id": {
            "example": "7c0b4a30-2f1d-4b3c-8e9a-1a2b3c4d5e6f",
            "format": "uuid",
            "type": "string"
          },
          "proposalId": {
            "example": "a1b2c3d4-e5f6-7890-abcd-ef0123456789",
            "format": "uuid",
            "type": "string"
          },
          "raisedBy": {
            "$ref": "#/components/schemas/ObjectionRaisedByDto"
          },
          "raisedById": {
            "example": "usr_3K9p1y2Mn8oW5dC7eUj4tZ",
            "type": "string"
          },
          "reason": {
            "example": "Capacity is too tight if we keep the plus-ones list as is.",
            "nullable": true,
            "type": "string"
          },
          "state": {
            "enum": [
              "live",
              "withdrawn"
            ],
            "example": "live",
            "type": "string"
          },
          "updatedAt": {
            "example": "2026-05-27T09:01:14.000Z",
            "type": "string"
          }
        },
        "required": [
          "id",
          "proposalId",
          "raisedById",
          "raisedBy",
          "state",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "PaymentDto": {
        "properties": {
          "amountMinor": {
            "description": "Amount in kobo (NGN minor unit) as a string — BigInt-safe.",
            "example": "3900000",
            "type": "string"
          },
          "currency": {
            "example": "NGN",
            "type": "string"
          },
          "eventId": {
            "example": "6bb69700-72eb-40a0-b37e-e6c0b8b004a9",
            "format": "uuid",
            "type": "string"
          },
          "failureReason": {
            "description": "Set when status=failed (e.g. \"card_declined\").",
            "example": null,
            "nullable": true,
            "type": "string"
          },
          "finalizedAt": {
            "example": "2026-05-26T08:14:23.000Z",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "example": "7a0a1b50-31c2-4d97-9d7e-1a5e0f0a2c4d",
            "format": "uuid",
            "type": "string"
          },
          "initiatedAt": {
            "example": "2026-05-26T08:12:00.000Z",
            "type": "string"
          },
          "processor": {
            "enum": [
              "paystack",
              "internal_free"
            ],
            "example": "paystack",
            "type": "string"
          },
          "processorReference": {
            "description": "Paystack reference (UUID) — null for the free-tier internal processor.",
            "example": "ref_3f8c2e1a4b9d",
            "nullable": true,
            "type": "string"
          },
          "purpose": {
            "description": "What the payment is for.",
            "enum": [
              "tier",
              "addon"
            ],
            "example": "tier",
            "type": "string"
          },
          "purposeRef": {
            "description": "Slug of the tier or add-on this payment unlocks.",
            "example": "occasion",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "success",
              "failed",
              "refunded"
            ],
            "example": "pending",
            "type": "string"
          },
          "taxMinor": {
            "description": "Nigerian VAT in kobo as a string — BigInt-safe.",
            "example": "292500",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "purpose",
          "amountMinor",
          "taxMinor",
          "currency",
          "processor",
          "status",
          "initiatedAt"
        ],
        "type": "object"
      },
      "PaymentIntentResponseDto": {
        "properties": {
          "amountMinor": {
            "description": "Total amount in kobo as a string — BigInt-safe.",
            "example": "3900000",
            "type": "string"
          },
          "authorizationUrl": {
            "description": "Paystack-hosted checkout URL. Null for free-tier (auto-success) or replays — the client should poll payment status instead of redirecting.",
            "example": "https://checkout.paystack.com/abc123",
            "nullable": true,
            "type": "string"
          },
          "currency": {
            "example": "NGN",
            "type": "string"
          },
          "paymentId": {
            "example": "7a0a1b50-31c2-4d97-9d7e-1a5e0f0a2c4d",
            "format": "uuid",
            "type": "string"
          },
          "reference": {
            "description": "Paystack reference — null for the free-tier internal processor.",
            "example": "ref_3f8c2e1a4b9d",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "success",
              "failed"
            ],
            "example": "pending",
            "type": "string"
          }
        },
        "required": [
          "paymentId",
          "status",
          "amountMinor",
          "currency"
        ],
        "type": "object"
      },
...
      "ReviewGivenDto": {
        "properties": {
          "comment": {
            "description": "Free-form comment. Surfaced only when the viewer is the reviewee or has admin rights on the event — null otherwise.",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-25T10:14:00.000Z",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/EventRefDto"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "revieweeId": {
            "example": "usr_creator...",
            "type": "string"
          },
          "reviewerId": {
            "example": "usr_2N9k8x...",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "reviewerId",
          "revieweeId",
          "rating",
          "createdAt",
          "event"
        ],
        "type": "object"
      },
      "ReviewReceivedDto": {
        "properties": {
          "comment": {
            "description": "Free-form comment. Surfaced only when the viewer is the reviewee or has admin rights on the event — null otherwise.",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-06-25T10:14:00.000Z",
            "type": "string"
          },
          "event": {
            "$ref": "#/components/schemas/EventRefDto"
          },
          "eventId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "rating": {
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          },
          "revieweeId": {
            "example": "usr_creator...",
            "type": "string"
          },
          "reviewer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewerSummaryDto"
              }
            ],
            "nullable": true
          },
          "reviewerId": {
            "example": "usr_2N9k8x...",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventId",
          "reviewerId",
          "revieweeId",
          "rating",
          "createdAt",
          "event"
        ],
        "type": "object"
      },
...
      "SubmitReviewDto": {
        "properties": {
          "comment": {
            "description": "Optional free-form comment. Only visible to the event creator (visibility rule per Phase 7 plan — public surface shows the aggregate only).",
            "example": "Tunde organised every segment crisply — the proposal-objection rhythm meant nobody felt steamrolled.",
            "maxLength": 2000,
            "type": "string"
          },
          "rating": {
            "description": "1-5 stars.",
            "example": 5,
            "maximum": 5,
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "rating"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "cookie": {
        "in": "cookie",
        "name": "better-auth.session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Event Planning Decision Making Assistant. Spec frozen per Phase 5.5 — bump `info.version` (major for breaking, minor for additive) on every committed change.",
    "title": "EPDEMA API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/events/{eventId}/reviews": {
      "get": {
        "operationId": "reviews_listForEvent",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReviewsListResponseDto"
                }
              }
            },
            "description": "Admin or creator gets every review with reviewer + comment hydrated. A non-privileged planning member sees only their own row. Non-members get 403 from the assertMember guard."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reviews on this event (admin/creator see all; member sees own)",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "reviews_submit",
        "parameters": [
          {
            "in": "path",
            "name": "eventId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReviewDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventReviewDto"
                }
              }
            },
            "description": "Submit one review per planning member per event. Event must be in state `past`. Reviewer cannot review their own event. 409 on a second submission for the same (reviewer, event). Atomic with the rolling average update on the creator's `user_profile.ratingAvg`."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a 1-5 review of the event creator (planning member only)",
        "tags": [
          "reviews"
        ]
      }
    },
...
    "/objections/{objectionId}/withdraw": {
      "post": {
        "operationId": "objections_withdraw",
        "parameters": [
          {
            "description": "Objection UUID.",
            "in": "path",
            "name": "objectionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectionResponseDto"
                }
              }
            },
            "description": "Withdraw an objection (objector only)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw an objection (objector)",
        "tags": [
          "objections"
        ]
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "payments_get",
        "parameters": [
          {
            "description": "Payment UUID.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "When \"true\" or \"1\", calls Paystack verify if the payment is still pending. Use on the return page.",
            "in": "query",
            "name": "reconcile",
            "required": false,
            "schema": {
              "example": "true",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            },
            "description": "Get a payment owned by the authenticated user."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a payment by ID",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/by-reference/{reference}": {
      "get": {
        "operationId": "payments_getByReference",
        "parameters": [
          {
            "description": "Paystack reference echoed back on the return-page redirect.",
            "in": "path",
            "name": "reference",
            "required": true,
            "schema": {
              "example": "ref_3f8c2e1a4b9d",
              "type": "string"
            }
          },
          {
            "description": "When \"true\" or \"1\", calls Paystack verify if the payment is still pending. Use on the return page.",
            "in": "query",
            "name": "reconcile",
            "required": false,
            "schema": {
              "example": "true",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDto"
                }
              }
            },
            "description": "Look up a payment by Paystack reference (used by the return page since Paystack only echoes `reference` in the redirect)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Look up a payment by Paystack reference",
        "tags": [
          "payments"
        ]
      }
    },
    "/payments/intents": {
      "post": {
        "operationId": "payments_createIntent",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "UUID v4 generated client-side.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentIntentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntentResponseDto"
                }
              }
            },
            "description": "Created (or replayed) a payment intent. Returns a Paystack checkout URL for paid tiers; free tier short-circuits to status=success with no URL."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create (or replay) a payment intent",
        "tags": [
          "payments"
        ]
      }
    },
    "/proposals/{proposalId}/objections": {
      "get": {
        "operationId": "objections_list",
        "parameters": [
          {
            "description": "Parent proposal UUID.",
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ObjectionResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List live objections against a proposal."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List live objections on a proposal",
        "tags": [
          "objections"
        ]
      },
      "post": {
        "operationId": "objections_create",
        "parameters": [
          {
            "description": "Parent proposal UUID.",
            "in": "path",
            "name": "proposalId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateObjectionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectionResponseDto"
                }
              }
            },
            "description": "Raise an objection against a proposal."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Raise an objection on a proposal",
        "tags": [
          "objections"
        ]
      }
    },
...
    "/users/me/reviews-given": {
      "get": {
        "operationId": "reviews_listGivenByMe",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewGivenDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reviews I have submitted",
        "tags": [
          "reviews"
        ]
      }
    },
    "/users/me/reviews-received": {
      "get": {
        "operationId": "reviews_listReceivedByMe",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewReceivedDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reviews of my events (creator surface; comments included for self)",
        "tags": [
          "reviews"
        ]
      }
    }
  },
  "servers": [],
...
}